feat: Update favicon and page title
This commit updates the favicon and page title in the index.html file. The favicon has been changed to use the logo.ico file instead of the vite.svg file. The page title has been changed to "Globo Store" instead of "Vite + React + TS". These changes improve the branding and visual identity of the website.